refactor: 1ファイル1宣言に分割し、ファイル名と中身を一致させる
SuggestionCategory+PlaceSearch.swift に無関係な型を積み上げてしまっていた。
ファイル名が中身の一部しか説明しておらず、#36 で ko/es/fr のキーワードを
追加する際に辿り着けない。

## 分割前

SuggestionCategory+PlaceSearch.swift（4宣言）
  extension SuggestionCategory  ← ファイル名どおり
  enum RecommendationQueries    ← 目的地検索シート用。無関係
  struct LocalizedQuerySet      ← 共通の型
  extension Locale              ← 無関係
MKCoordinateRegion+Locale.swift（2宣言）
  extension MKCoordinateRegion
  extension Locale              ← 無関係

extension Locale が2ファイルに分散し、しかもどちらもファイル名と無関係な
場所にあった。「端末のロケールを取る処理」を探すのに2箇所見る必要があった。

## 分割後（すべて1宣言）

Models/SuggestionCategory+PlaceSearch.swift   extension SuggestionCategory
Models/RecommendationQueries.swift            enum RecommendationQueries
Utilities/LocalizedQuerySet.swift             struct LocalizedQuerySet
Utilities/Extensions/Locale+Current.swift     extension Locale（2つを集約）
Utilities/Extensions/MKCoordinateRegion+Locale.swift  extension MKCoordinateRegion

Locale+Current.swift には iOS 16 での API 変更（languageCode/regionCode の
非推奨化）を集約する意図もコメントに残した。

ロジックの変更は無し。テスト180件パス。
